requirements.txt
# Framework principal
fastapi==0.110.0

# Servidor ASGI
uvicorn[standard]==0.29.0

# Banco de dados e ORM
SQLAlchemy==2.0.30
alembic==1.13.1          # migrações

# Validação de dados
pydantic==2.7.0

# Autenticação e segurança
python-jose==3.3.0       # JWT
passlib[bcrypt]==1.7.4   # hash de senhas e bcrypt
bcrypt==4.1.2            # suporte extra de hash

# Utilitários
python-dotenv==1.0.0     # variáveis de ambiente
requests==2.31.0         # para testes
python-multipart==0.0.20 # para uploads de formulário no FastAPI
psutil==7.1.0            # para monitoramento do sistema

# Testes automatizados
selenium==4.15.2
webdriver-manager==4.0.1

#instalação de dependências
#instalação de dependências
#pip install -r requirements.txt

#rodar o servidor
#uvicorn app.main:app --reload